Added code to git
The pi_calculator Verilog file contains the top level module. The testbench
tests it for Verilog interpreters like QuestaSim. The sim_main.cpp file tests
it for Verilator (which is much faster than QuestaSim). The build.sh file
builds it for Verilator. The digit_tester.py file compares the transcript.txt
file with the pi.txt file (pi.txt contains the real digits). The
compile_flags.txt file is for clangd to not give errors when parsing the C++
file.